[FIX] web_services: fix db dump/restore when authentication requires a password on non-win32 systems
The various pg_* utilities require the password
via a special environment variable or a special
`pgpass` file in the user home, even on Unix,
namely when the PostgreSQL connection is done
via TCP and not via a unix socket.

Setting the environment variable is relatively
safe if it is removed from the environment
immediately after the operation, and saves user
the trouble of managing the pgpass file themselves.

This had been fixed at revision 3992 but was
incorrectly removed for non win32 platforms
at revision 4424 (rev-id stw@openerp.com-20120912114651-8hcliparft1ep9tc)